Refactor. Moved parts of back-reference.* out of tools to main folder.

+++ tools/lz77.c
@@ -1,5 +1,5 @@
 #include "lz77.h"
-#include "lz77-back-reference.h"
+#include "lz77-back-reference.tools.h"
 
 void lz77Compress(Buffer *uncompressed, Buffer *compressed)
 {
@@ -18,7 +18,7 @@ void lz77Compress(Buffer *uncompressed, Buffer *compressed)
 		}
 
 		// Write a back-reference.
-		BackReference backReference = backReferenceFind(uncompressed, read - uncompressed->content);
+		BackReference backReference = backReferenceFind(uncompressed, (uint16_t)(read - uncompressed->content));
 		backReferenceEncodeToBuffer(&backReference, compressed);
 		read += backReference.length;
